requirements.txt
streamlit
openai
python-dotenv
PyMuPDF
langchain
faiss-cpu
